Move App code into RouterProvider
