+++ frontend/app-development/layout/Layout.module.css
@@ -7,7 +7,7 @@
   );
   --left-menu-width: 68px;
 
-  background-color: lightgray;
+  background: white;
   min-height: 100vh;
   width: 100%;
   display: flex;
