+++ frontend/app-development/layout/Layout.test.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
-import { App } from './App';
-import type { IUserState } from './sharedResources/user/userSlice';
+import { Layout } from './Layout';
+import type { IUserState } from '../sharedResources/user/userSlice';
 import { screen } from '@testing-library/react';
 import { APP_DEVELOPMENT_BASENAME } from 'app-shared/constants';
-import { renderWithProviders } from './test/testUtils';
-import * as testids from '../testing/testids';
-import { textMock } from '../testing/mocks/i18nMock';
+import { renderWithProviders } from '../test/testUtils';
+import * as testids from '../../testing/testids';
+import { textMock } from '../../testing/mocks/i18nMock';
 import { queriesMock } from 'app-shared/mocks/queriesMock';
 
 jest.mock('../language/src/nb.json', jest.fn());
@@ -14,7 +14,7 @@ jest.mock('../language/src/en.json', jest.fn());
 // Mocking console.error due to Tanstack Query removing custom logger between V4 and v5 see issue: #11692
 const realConsole = console;
 const render = async (remainingMinutes: number = 40) => {
-  renderWithProviders(<App />, {
+  renderWithProviders(<Layout />, {
     startUrl: `${APP_DEVELOPMENT_BASENAME}/my-org/my-app`,
     queries: { ...queriesMock },
     preloadedState: {
@@ -26,7 +26,7 @@ const render = async (remainingMinutes: number = 40) => {
     },
   });
 };
-describe('App', () => {
+describe('Layout', () => {
   beforeEach(() => {
     global.console = {
       ...console,
