--- frontend/app-development/router/PageRoutes.module.css
@@ -1,6 +1,0 @@
-.root {
-  display: flex;
-  flex-direction: column;
-  min-height: 100vh;
-  background: white;
-}
